An origin and destination table of workers, place of employment, trip counts, and estimated travel times.

This object contains the trips made from origin (TAZ) to destination (TAZ) by full-time workers to work destinations; it is sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area. Also contained within is the calculated car travel time for each trip calculated using TAZ centroids and r5r in October 2021.

Usage

data(od)

Format

A dataframe containing 103076 rows and 17 variables; each row represents a unique trip from origin TAZ to destination TAZ with associated features.

Origin

The unique ID of the origin traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GTA06'.

Destination

The unique ID of the destination traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GTA06'.

Persons

The number of full-time workers at the origin who work at the destination.

T_LT

The number of work-related trips taken from the origin to the destination by local transit excluding GO transit (regional transit service) as a primary mode.

T_LTGO

The number of work-related trips taken from the origin to the destination by local transit and GO transit (regional transit service) as a primary mode.

T_GO

The number of work-related trips taken from the origin to the destination by GO transit (regional transit service as a primary mode.

BC

The number of work-related trips taken from the origin to the destination by bicyle (primary mode).

W

The number of work-related trips taken from the origin to the destination by walking (primary mode).

C_DPC

The number of work-related trips taken from the origin to the destination as a driver of a private car (primary mode).

C_PPC

The number of work-related trips taken from the origin to the destination as a passenger of a private car (primary mode).

C_TA

The number of work-related trips taken from the origin to the destination by a taxi (primary mode).

C_RS

The number of work-related trips taken from the origin to the destination by a paid ride-share (primary mode).

C_SB

The number of work-related trips taken from the origin to the destination by a school bus (primary mode).

M

The number of work-related trips taken from the origin to the destination by a motorcycle (primary mode).

O

The number of work-related trips taken from the origin to the destination by another mode (primary mode).

sum_trips

The total number of work-related trips taken from the origin to the destination by all modes.

travel_time

The car travel time from origin to destination assuming a 7:00 am departure from and to the TAZ centroids on Oct. 20 2021.

Source

"2016 Transportation Tomorrow Survey" from the Data Management Group accessed November 14 2021.

Travel times calculated using r5r